+++ implement/Pine.Core/Expression.cs
@@ -1,7 +1,6 @@
 using Pine.Json;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json.Serialization;
 
 namespace Pine.Core;