+++ implement/pine/JavaScript/JavaScriptEngineJintOptimizedForElmApps.cs
@@ -3,7 +3,6 @@ using Jint.Native;
 using Jint.Native.Object;
 using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Text;