+++ implement/test-elm-time/AVH4ElmFormatBinariesTests.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Pine.Elm;
-using System.Text;
 
 namespace TestElmTime;