+++ implement/test-elm-time/ParseElmInteractiveSubmissionTests.cs
@@ -21,6 +21,11 @@ public class ParseElmInteractiveSubmissionTests
             },
             new
             {
+                SubmissionText = "test",
+                ExpectedResponse = "ExpressionSubmission (FunctionOrValue [] \"test\")"
+            },
+            new
+            {
                 SubmissionText = "greet \"World\"",
                 ExpectedResponse = "ExpressionSubmission (Application [Node { end = { column = 10, row = 7 }, start = { column = 5, row = 7 } } (FunctionOrValue [] \"greet\"),Node { end = { column = 18, row = 7 }, start = { column = 11, row = 7 } } (Literal \"World\")])"
             },
@@ -29,6 +34,11 @@ public class ParseElmInteractiveSubmissionTests
                 SubmissionText = "replicate 2 3",
                 ExpectedResponse = "ExpressionSubmission (Application [Node { end = { column = 14, row = 7 }, start = { column = 5, row = 7 } } (FunctionOrValue [] \"replicate\"),Node { end = { column = 16, row = 7 }, start = { column = 15, row = 7 } } (Integer 2),Node { end = { column = 18, row = 7 }, start = { column = 17, row = 7 } } (Integer 3)])"
             },
+            new
+            {
+                SubmissionText = "test = 123",
+                ExpectedResponse = "DeclarationSubmission (FunctionDeclaration { declaration = Node { end = { column = 11, row = 6 }, start = { column = 1, row = 6 } } { arguments = [], expression = Node { end = { column = 11, row = 6 }, start = { column = 8, row = 6 } } (Integer 123), name = Node { end = { column = 5, row = 6 }, start = { column = 1, row = 6 } } \"test\" }, documentation = Nothing, signature = Nothing })"
+            }
         };
 
         for (var i = 0; i < testCases.Length; i++)